+++ CHANGES.md
@@ -71,6 +71,8 @@ Unreleased
   runtimes ([#63](https://github.com/anmonteiro/httpaf/pull/63))
 - httpaf-async: Use [gluten](https://github.com/anmonteiro/gluten) to implement
   the Async runtimes ([#69](https://github.com/anmonteiro/httpaf/pull/69))
+- httpaf: call error handler on read EOF if the entire body hasn't been
+  received ([#75](https://github.com/anmonteiro/httpaf/pull/75))
 
 httpaf (upstream) 0.6.6
 --------------